% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_AT_url.R
\name{build_AT_url}
\alias{build_AT_url}
\title{Create URL's for downloading AT data from the cloud}
\usage{
build_AT_url(x, dates = as.character(Sys.Date()), limit = 1000)
}
\arguments{
\item{x}{data frame with cols for "key" and "mac"}

\item{dates}{what dates to retrieve format is, e.g., "2022-10-21"}

\item{limit}{maximum number of records to return per date}
}
\value{
a vector of character strings containing urls
}
\description{
Create API URL's for retrieving AtmoTube data
}
\author{
Neil E. Klepeis
}
