% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_AT_cloud_data.R
\name{get_AT_cloud_data}
\alias{get_AT_cloud_data}
\title{Get AtmoTube Sensor Data via Cloud API}
\usage{
get_AT_cloud_data(
  key,
  mac,
  limit = 1000,
  offset = 0,
  format = "csv",
  order = "desc",
  date = NULL,
  ...
)
}
\arguments{
\item{key}{string containing API access key (obtained from AtmoTube)}

\item{mac}{MAC address for the device}

\item{limit}{maximum number of records to return}

\item{offset}{skip records}

\item{format}{"csv" or "json"}

\item{order}{"asc" or "desc" time order of records}

\item{date}{filter by date, e.g., "2022-10-21"}

\item{...}{additional arguments passed to "fromJSON" function}
}
\value{
a tibble containing rectangular AtmoTube data with one
record per time point and multiple columns containing sensor data
streams (\code{format="csv"}) OR a list representing a JSON object containing
the sensor data and various flags (\code{format="json"})
}
\description{
Retrieve AtmoTube Air Monitor sensor data using the Cloud API
}
\details{
The AtmoTube Cloud API contains features to retrieve
air sensor data for a single monitor with a specified limit to the number
of records and the time-order of the records.

For the AtmoTube PRO, sensor data streams include particulate matter in
PM1, PM2.5, and PM10 ranges (ug/m3), VOC's (ppm),
temperature (degrees C), humidity (\%), atmospheric pressure (mbar),
latitude, and longitude.
\subsection{Parameters}{
\itemize{
\item mac - MAC address of Atmotube device
\item order - Time descending "desc" or ascending "asc"
\item format - "json" or "csv"
\item offset - records to skip for pagination, min of 0
\item date - date filter, enter date, e.g., "2022-10-1"
\item limit - maximum number of records to return, up to 1000
}
}
}
\examples{

key <- "190b3980-d505-11eb-b8bc-0242ac130003"
mac <- "DF:21:2A:AE:9A:17"
x <- get_AT_cloud_data(key, mac)

# https://api.atmotube.com/api/v1/data?api_key=190b3980-d505-11eb-b8bc-0242ac130003&mac=DF:21:2A:AE:9A:17&limit=199999&offset=0&order=desc

}
\author{
Neil E. Klepeis
}
