% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_AT_cloud_data_multi.R
\name{get_AT_cloud_data_multi}
\alias{get_AT_cloud_data_multi}
\title{Get AtmoTube Sensor Data via Cloud API for Multiple Devices
Over a range of specified days}
\usage{
get_AT_cloud_data_multi(
  key,
  macs,
  dates = as.character(Sys.Date()),
  limit = 1000,
  format = "csv"
)
}
\arguments{
\item{key}{string containing API access key (obtained from AtmoTube)}

\item{macs}{MAC addresses for the devices}

\item{dates}{dates for which to retrieve data}

\item{limit}{maximum number of records to return, max 1000}
}
\value{
a tibble containing rectangular AtmoTube data with one
record per time point and multiple columns containing sensor data
streams
}
\description{
Retrieve AtmoTube Air Monitor sensor data using the Cloud API
}
\details{
The AtmoTube Cloud API contains features to retrieve
air sensor data in CSV format for multiple monitors with a specified limit to the number
of records and the time-order of the records.

Sets offset to 0 and order to "asc".

For the AtmoTube PRO, sensor data streams include particulate matter in
PM1, PM2.5, and PM10 ranges (ug/m3), VOC's (ppm),
temperature (degrees C), humidity (\%), atmospheric pressure (mbar),
latitude, and longitude.
\subsection{Parameters}{
\itemize{
\item macs - MAC addresses of 1 or more Atmotube devices
\item order - Time descending "desc" or ascending "asc"
\item offset - records to skip for pagination, min of 0
\item date - date filter, enter date, e.g., "2022-10-1"
\item limit - maximum number of records to return, up to 1000
}
}
}
\examples{

key <- "190b3980-d505-11eb-b8bc-0242ac130003"
mac <- "DF:21:2A:AE:9A:17"
x <- get_AT_cloud_data(key, mac)

# https://api.atmotube.com/api/v1/data?api_key=190b3980-d505-11eb-b8bc-0242ac130003&mac=DF:21:2A:AE:9A:17&limit=199999&offset=0&order=desc

}
\author{
Neil E. Klepeis
}
