% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_AT_from_device_list.R
\name{get_AT_from_device_list}
\alias{get_AT_from_device_list}
\title{Get Daily AtmoTube Sensor Data Based on a list of devices}
\usage{
get_AT_from_device_list(devices, start_date, end_date)
}
\arguments{
\item{devices}{a data from containing cols for "mac"=MacID, "key"=Access Key,
"did"=device ID, and "pid" = Participant or Person ID.}

\item{start_date}{staring containg starting date in "YYYY-MM-DD" format}

\item{end_date}{string containg ending date in "YYYY-MM-DD" format}
}
\value{
a tibble containing rectangular AtmoTube data with one
record per time point and multiple columns containing sensor data
streams
}
\description{
Retrieve daily AtmoTube Air Monitor sensor data using a list
of devices and their associated MacID and Access Key
}
\details{
This function is a front-end to get_AT_cloud_data_multi
but takes a list of device and conveniently returns data for one or
more devices for every date in between 'start' and 'end' dates.
}
\author{
Neil E. Klepeis
}
