% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_AT_device_summary.R
\name{make_AT_device_summary}
\alias{make_AT_device_summary}
\title{Make a summary of AtmoTube data}
\usage{
make_AT_device_summary(data)
}
\arguments{
\item{data}{a data frame with AtmoTube data merged with device list
data}
}
\value{
a tibble
}
\description{
Create a statistical summary of AtmoTube data merged
with device list data
}
\author{
Neil E. Klepeis
}
