% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_AT_devices.R
\name{plot_AT_devices}
\alias{plot_AT_devices}
\title{Plot AtmoTube data merged with device info}
\usage{
plot_AT_devices(
  data,
  cols = "PM2.5 (ug/m3)",
  id.cols = "Mac",
  tz = "America/Los_Angeles",
  by = "1 day",
  date_breaks = "6 hour"
)
}
\arguments{
\item{data}{a data frame with AtmoTube data merged with device list
data}
}
\value{
a tibble
}
\description{
Create a ggplot object of AtmoTube data merged with device info
}
\author{
Neil E. Klepeis
}
