% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_AT_map_ggplot.R
\name{plot_AT_map_ggplot}
\alias{plot_AT_map_ggplot}
\title{Plot AtmoTube trajectory on a map}
\usage{
plot_AT_map_ggplot(
  data,
  by = "1 day",
  streets = TRUE,
  place = TRUE,
  water = TRUE,
  rail = TRUE,
  building = TRUE
)
}
\arguments{
\item{data}{a tibble of AtmoTube data from Cloud API}
}
\description{
Plot AtmoTube trajectory on a map
}
\details{
This function plots a static geographic map showing where the
AtmoTube was located by day/hour/etc

Uses the osmdata (Open Stream Map) package and ggplot to create
the plot.
}
\author{
Neil E. Klepeis
}
