% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_AT_map_leaflet.R
\name{plot_AT_map_leaflet}
\alias{plot_AT_map_leaflet}
\title{Plot AtmoTube trajectory on a interactive map using leaflet}
\usage{
plot_AT_map_leaflet(data)
}
\arguments{
\item{data}{a tibble of AtmoTube data from Cloud API}
}
\description{
Plot AtmoTube trajectory on a interactive map using leaflet
}
\details{
This function plots an interactive map showing where the
AtmoTube was located by day/hour/etc
}
\author{
Neil E. Klepeis
}
