% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_AT_streams.R
\name{plot_AT_streams}
\alias{plot_AT_streams}
\title{Plot AtmoTube sensor streams}
\usage{
plot_AT_streams(
  data,
  cols = c("PM1,", "PM2.5,", "PM10,", "VOC"),
  id.cols = "Mac",
  time.col = "Date (GMT)",
  by = "1 day",
  tz = "America/Los_Angeles",
  ...
)
}
\arguments{
\item{data}{a tibble of AtmoTube data from Cloud API}

\item{cols}{character vector of columns names corresponding to sensor response streams to plot.
Names of responses in the columns to plot, defaults to PM and VOC columns.}

\item{id.cols}{names of device-ID columns to use in faceting the plot,
e.g., Mac address or site name., defaults to "Mac"}

\item{time.col}{name of column containing the time}

\item{by, }{string giving the time window for faceting, e.g., "1 day", "1 month",
default to "1 day"}

\item{tz}{the time zone}
}
\description{
Plot streams downloaded using the AtmoTube API
}
\details{
Converts a raw wide file downloaded using the AtmoTube API to a
standard long format with Time, Response, and Value columns.  id.cols contains
extra index columns to use in faceting, defaults to "Mac" for the device
MAC address.
}
\author{
Neil E. Klepeis
}
